The O class file that modded kik originally has
It is in a different file location. The o.class {(o.java, o.smali)} file is suppost to be in the directory path: ...\kik\a\b\"o.class", however, it does not. Supposedly, the "o class" files moves with updates, so that must be the case. Looking at the file, it seems true, but the imports are different.